feat: DF-380 - Resolve circular dependency
